shuffling files around in the probe, before improving it
