+++ nym-gateway-probe/src/common/bandwidth_helpers.rs
